[add]better UI for HomeAppBar
