+++ .idea/deploymentTargetDropDown.xml
@@ -2,6 +2,12 @@
 <project version="4">
   <component name="deploymentTargetDropDown">
     <value>
+      <entry key="HomeAppBarPreview">
+        <State />
+      </entry>
+      <entry key="HomeScreenPreview">
+        <State />
+      </entry>
       <entry key="app">
         <State />
       </entry>
