+++ app/src/main/java/com/example/okura/presentation/view/HomeScreen.kt
@@ -1,9 +1,88 @@
 package com.example.okura.presentation.view
 
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AccountCircle
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.NavHostController
+import androidx.navigation.compose.rememberNavController
+import com.example.okura.ui.theme.OkuraTheme
 
 @Composable
 fun HomeScreen(navController: NavHostController) {
+    Scaffold(
+        topBar = {
+            HomeAppBar(
+                title = { Text(text = "Okura") })
+        }
+    ) { paddingValues ->
+        Text(text = "Home", modifier = Modifier.padding(paddingValues))
+    }
+}
 
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun HomeAppBar(
+    title: @Composable () -> Unit,
+    modifier: Modifier = Modifier,
+    navigationIcon: @Composable (() -> Unit)? = null,
+    actions: @Composable RowScope.() -> Unit = {}
+) {
+    val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior(rememberTopAppBarState())
+
+    CenterAlignedTopAppBar(
+        title = { title() },
+        navigationIcon = {
+            IconButton(onClick = { /* do something */ }) {
+                Icon(
+                    imageVector = Icons.Filled.Favorite,
+                    contentDescription = "Localized description"
+                )
+            }
+        },
+        actions = {
+            IconButton(
+                onClick = { /*TODO*/ }
+            ) {
+                Icon(
+                    imageVector = Icons.Default.AccountCircle,
+                    contentDescription = "Profile"
+                )
+            }
+        },
+        scrollBehavior = scrollBehavior,
+        colors = TopAppBarDefaults.topAppBarColors(
+            containerColor = Color.Transparent,
+            titleContentColor = MaterialTheme.colorScheme.primary
+        ),
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun HomeScreenPreview() {
+    OkuraTheme {
+        val navController = rememberNavController() // dummy navController for preview
+        HomeScreen(navController = navController)
+    }
+}
+
+@Preview
+@Composable
+fun HomeAppBarPreview() {
+    HomeAppBar(title = { /*TODO*/ })
 }
